Modificacion

+++ src/main/java/pe/edu/upc/donfy/repositories/IRoleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upc.donfy.entities.Role;
 
 @Repository
-public interface IRoleRepository extends JpaRepository<Role, Integer> {
+public interface IRoleRepository extends JpaRepository<Role, Long> {
 }
